requirements.txt
alembic==1.17.2
asyncpg==0.31.0
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.127.0
passlib==1.7.4
pydantic-settings==2.12.0
SQLAlchemy==2.0.45
uvicorn==0.40.0
redis==7.1.0
python-jose==3.5.0